% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{deprecated_functions}
\alias{deprecated_functions}
\alias{osm_poly}
\alias{osm_line}
\alias{osm_point}
\title{Deprecated tmaptools functions}
\usage{
osm_poly()

osm_line()

osm_point()
}
\description{
The following functions are not used anymore or deprecated as of tmaptools version 2.0
}
\details{
\itemize{
 \item \code{osm_poly}: for reading vector OSM data, use the \code{osmdata} package
 \item \code{osm_line}: for reading vector OSM data, use the \code{osmdata} package
 \item \code{osm_point}: for reading vector OSM data, use the \code{osmdata} package
}
}
\keyword{internal}
