% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_brewer_pal.R
\name{get_brewer_pal}
\alias{get_brewer_pal}
\title{Get and plot a (modified) Color Brewer palette}
\usage{
get_brewer_pal(palette, n = 5, contrast = NA, stretch = TRUE, plot = TRUE)
}
\arguments{
\item{palette}{name of the color brewer palette. Run \code{\link{palette_explorer}} or see \code{\link[RColorBrewer:brewer.pal]{brewer.pal}} for options.}

\item{n}{number of colors}

\item{contrast}{a vector of two numbers between 0 and 1 that defines the contrast range of the palette. Applicable to sequential and diverging palettes. For sequential palettes, 0 stands for the leftmost color and 1 the rightmost color. For instance, when \code{contrast=c(.25, .75)}, then the palette ranges from 1/4 to 3/4 of the available color range. For diverging palettes, 0 stands for the middle color and 1 for both outer colors. If only one number is provided, the other number is set to 0. The default value depends on \code{n}. See details.}

\item{stretch}{logical that determines whether intermediate colors are used for a categorical palette when \code{n} is greater than the number of available colors.}

\item{plot}{should the palette be plot, or only returned? If \code{TRUE} the palette is silently returned.}
}
\value{
vector of color values. It is silently returned when \code{plot=TRUE}.
}
\description{
Get and plot a (modified) palette from Color Brewer. In addition to the base function \code{\link[RColorBrewer:brewer.pal]{brewer.pal}}, a palette can be created for any number of classes. The contrast of the palette can be adjusted for sequential and diverging palettes. For categorical palettes, intermediate colors can be generated. An interactive tool that uses this function is \code{\link{palette_explorer}}.
}
\details{
The default contrast of the palette depends on the number of colors, \code{n}, in the following way. The default contrast is maximal, so \code{(0, 1)}, when \code{n = 9} for sequential palettes and \code{n = 11} for diverging palettes. The default contrast values for smaller values of \code{n} can be extracted with some R magic: \code{sapply(1:9, tmaptools:::default_contrast_seq)} for sequential palettes and \code{sapply(1:11, tmaptools:::default_contrast_div)} for diverging palettes.
}
\examples{
get_brewer_pal("Blues")
get_brewer_pal("Blues", contrast=c(.4, .8))
get_brewer_pal("Blues", contrast=c(0, 1))
get_brewer_pal("Blues", n=15, contrast=c(0, 1))

get_brewer_pal("RdYlGn")
get_brewer_pal("RdYlGn", n=11)
get_brewer_pal("RdYlGn", n=11, contrast=c(0, .4))
get_brewer_pal("RdYlGn", n=11, contrast=c(.4, 1))

get_brewer_pal("Set2", n = 12)
get_brewer_pal("Set2", n = 12, stretch = FALSE)
}
\seealso{
\code{\link{palette_explorer}}
}
