% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_coloring.R
\name{get_neighbours}
\alias{get_neighbours}
\title{Get neighbours list from spatial objects}
\usage{
get_neighbours(x)
}
\arguments{
\item{x}{a shape object, i.e., a \code{sf} object or a \code{\link[sp:SpatialPolygonsDataFrame]{SpatialPolygons(DataFrame)}}.}
}
\value{
A list where the items correspond to the features. Each item is a vector of neighbours.
}
\description{
Get neighbours list from spatial objects. The output is similar to the function \code{poly2nb} of the \code{spdep} package, but uses \code{sf} instead of \code{sp}.
}
