% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_shape_units.R
\name{get_shape_units}
\alias{get_shape_units}
\title{Get shape units}
\usage{
get_shape_units(x = NULL, projection = NULL, latitude = NULL,
  target.unit = NULL)
}
\arguments{
\item{x}{a shape object from class Spatial, Raster or sf}

\item{projection}{projection string}

\item{latitude}{latitude, in case \code{x} is unprojected}

\item{target.unit}{target unit, one of "km", "m", "mi", and "ft"}
}
\value{
list with the following items, \code{unit} according to the coordinates/projection and \code{to_meter} which is the multiplier to get meters. If \code{target.unit} is specified, the list also contains \code{target} and \code{to_target}, which are respectively the target.unit and the multipler to get these units.
}
\description{
Get the units in which the shape coordinates are specified, and determine the multiplier to transpose it in meters or other units. For unprojected shapes, the conversion of coordinates to meters hold for a specific latitude.
}
\examples{
if (require(tmap)) {
    data(World, land)
    get_shape_units(World, target.unit = "miles")

    get_shape_units(land)
    get_shape_units(land, latitude = 45)
    get_shape_units(land, latitude = 89)
}
}
\seealso{
\code{\link{approx_distances}} for distances between two points
}

