% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette_explorer.R
\name{palette_explorer}
\alias{palette_explorer}
\title{Explore color palettes}
\usage{
palette_explorer()
}
\description{
This interactive tool shows all Color Brewer palettes, where the number of colors can be adjusted as well as the contrast range. Categorical (qualitative) palettes can be stretched when the number of colors exceeds the number of palette colors. Output code needed to get the desired color values is generated. Finally, all colors can be tested for color blindness.
}
\examples{
\dontrun{
if (require(shiny) && require(shinyjs)) {
    palette_explorer()
}
}
}
\references{
\url{http://www.color-blindness.com/types-of-color-blindness/}

Tennekes, M., 2018, {tmap}: Thematic Maps in {R}, Journal of Statistical Software, 84(6), 1-39, \href{https://doi.org/10.18637/jss.v084.i06}{DOI}
}
\seealso{
\code{\link{get_brewer_pal}}, \code{\link[dichromat:dichromat]{dichromat}}, \code{\link[RColorBrewer:RColorBrewer]{RColorBrewer}}
}
