% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmaptools-package.R
\name{\%>\%}
\alias{\%>\%}
\title{Pipe operator}
\arguments{
\item{lhs}{Left-hand side}

\item{rhs}{Right-hand side}
}
\description{
The pipe operator from magrittr, \code{\%>\%}, can also be used in functions from \code{tmaptools}.
}
