% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\docType{package}
\name{tmaptools-deprecated}
\alias{tmaptools-deprecated}
\alias{tmaptools-deprecated-package}
\title{Deprecated functions in \code{tmaptools}}
\description{
These functions are based on the \code{sp} package and are not supported anymore. They have been migrated to \url{https://github.com/mtennekes/oldtmaptools}
}
\details{
\code{\link{append_data}}, \code{\link{aggregate_map}}, \code{\link{double_line}}, \code{\link{points_to_raster}}, \code{\link{poly_to_raster}}, \code{\link{sample_dots}}, \code{\link{sbind}}, \code{\link{smooth_map}}, \code{\link{smooth_raster_cover}}, \code{\link{read_shape}}, \code{\link{write_shape}}
}
