% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{deprecated_functions}
\alias{deprecated_functions}
\alias{get_proj4}
\alias{set_projection}
\alias{get_projection}
\alias{is_projected}
\title{Deprecated tmaptools functions}
\usage{
get_proj4(x, output = c("crs", "character", "epsg", "CRS"))

set_projection(
  shp,
  projection = NA,
  current.projection = NA,
  overwrite.current.projection = FALSE
)

get_projection(
  shp,
  guess.longlat = FALSE,
  output = c("character", "crs", "epsg", "CRS")
)

is_projected(x)
}
\arguments{
\item{x}{see documentation in tmaptools 2.x for details}

\item{output}{see documentation in tmaptools 2.x for details}

\item{shp}{see documentation in tmaptools 2.x for details}

\item{projection}{see documentation in tmaptools 2.x for details}

\item{current.projection}{see documentation in tmaptools 2.x for details}

\item{overwrite.current.projection}{see documentation in tmaptools 2.x for details}

\item{guess.longlat}{see documentation in tmaptools 2.x for details

Deprecated as of version 2.0
\itemize{
 \item \code{append_data}, \code{aggregate_map}, \code{double_line}, \code{points_to_raster}, \code{poly_to_raster}, \code{sample_dots}, \code{sbind}, \code{smooth_map}, \code{smooth_raster_cover}, \code{read_shape}, \code{write_shape}. These functions are based on the \code{sp} package and are not supported anymore. They have been migrated to \url{https://github.com/mtennekes/oldtmaptools}
 \item \code{osm_poly}, \code{osm_line}, \code{osm_point}. Please use the package \code{osmdata}
}

Deprecated as of version 3.0
\itemize{
 \item \code{get_proj4}: for projections, please use the function \code{\link[sf:st_crs]{st_crs}}
 \item \code{set_projection} for setting map projections, use \code{\link[sf:st_crs]{st_crs}} or \code{\link[sf:st_transform]{st_transform}}
 \item \code{get_projection} for getting map projections, use \code{\link[sf:st_crs]{st_crs}}
 \item \code{is_projected} for checking if projections are long lat coordinates, use \code{\link[sf:st_is_longlat]{st_is_longlat}}
}}
}
\description{
The following functions are not used anymore or deprecated as of tmaptools version 3.0. These functions are based on the \code{sp} package and are not supported anymore. They have been migrated to \url{https://github.com/mtennekes/oldtmaptools}
}
\keyword{internal}
