% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\name{\%>\%}
\alias{\%>\%}
\title{Pipe operator}
\arguments{
\item{lhs}{Left-hand side}

\item{rhs}{Right-hand side}
}
\description{
The pipe operator from magrittr, \code{\%>\%}, can also be used in functions from \code{tmaptools}.
}
\examples{
if (require(tmap)) {
    data(land, World)

    current.mode <- tmap_mode("view")

    land \%>\%
        crop_shape(World[World$name=="China",], polygon = TRUE) \%>\%
        tm_shape() +
        tm_raster("cover_cls")

    tmap_mode(current.mode)
}
}

